requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.10
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
drf-yasg==1.21.11
gunicorn==23.0.0
inflection==0.5.1
kombu==5.6.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pdf2image==1.17.0
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sqlparse==0.5.5
tesseract==0.1.3
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
